requirements.txt
django==1.8.2
path.py
django-model-utils==2.2
gunicorn==19.3.0
dj-database-url==0.3.0
dj-static==0.0.6
psycopg2==2.6
django-celery==3.1.16
redis==2.10.3
django-registration-redux==1.2
django-celery-email==1.1.1
# The django_ses project has been taken over. Replace when release > 0.7.0 happens (see issue #70)
git+git://github.com/django-ses/django-ses.git@2fae0ebdaa96ff6daaea4edaafca3c02cdda61d8#egg=django_ses
django-crispy-forms==1.4.0
-e git+git://github.com/adamcharnock/django-easydump.git@master#egg=easydump
django-redis==4.0.0
django-pipeline==1.5.1
django-extensions==1.5.5
django_smalluuid==0.1.2
boto3==1.1.0
django-storages-redux==1.2.3

